+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ## 1.0.0
 
 * initial project.
+
+## 1.0.1
+
+* add android ios platform base features
+* add try catch for all methods
+* add example